fix: changed preparedStatement of registrationNumber in register

+++ proyectoClinica/src/main/java/com/example/proyectoClinica/repository/daos/impl/DentistDaoH2.java
@@ -24,7 +24,7 @@ public class DentistDaoH2 implements IDao<Dentist> {
 
             //Create sentence
             preparedStatement = connection.prepareStatement("INSERT INTO dentist (registrationNumber, firstName, lastName) VALUES(?, ?, ?)", Statement.RETURN_GENERATED_KEYS);
-            preparedStatement.setLong(1, dentist.getRegistrationNumber());
+            preparedStatement.setInt(1, dentist.getRegistrationNumber());
             preparedStatement.setString(2, dentist.getFirstName());
             preparedStatement.setString(3, dentist.getLastName());
 
@@ -147,15 +147,14 @@ public class DentistDaoH2 implements IDao<Dentist> {
     @Override
     public Dentist update(Dentist dentist) {
         try {
-            //2 Crear una sentencia especificando que el ID lo auto incrementa la base de datos y que nos devuelva esa Key es decir ID
+            //Create sentence
             preparedStatement = connection.prepareStatement("UPDATE address SET registrationNumber = ?, firstName = ?, lastName = ? WHERE id = ?");
-            //No le vamos a pasar el ID ya que hicimos que fuera autoincremental en la base de datos
             preparedStatement.setLong(1, dentist.getRegistrationNumber());
             preparedStatement.setString(2, dentist.getFirstName());
             preparedStatement.setString(3, dentist.getLastName());
             preparedStatement.setLong(4, dentist.getId());
 
-            //3 Ejecutar una sentencia SQL
+            //Execute a SQL sentence
             preparedStatement.executeUpdate();
 
             preparedStatement.close();
